+++ common/experiment/formats/documents.py
@@ -1,3 +1,6 @@
+from arekit.processing.text.parser import TextParser
+
+
 class DocumentOperations(object):
     """
     Provides operations with documents
@@ -15,12 +18,11 @@ class DocumentOperations(object):
     def iter_news_indices(self, data_type):
         raise NotImplementedError()
 
-    def iter_parsed_news(self, doc_inds, frame_variant_collection):
+    def iter_parsed_news(self, doc_inds):
         for doc_id in doc_inds:
-            yield self.__parse_news(doc_id=doc_id,
-                                    frame_variant_collection=frame_variant_collection)
+            yield self.__parse_news(doc_id=doc_id)
 
-    def __parse_news(self, doc_id, frame_variant_collection):
+    def __parse_news(self, doc_id):
         news = self.read_news(doc_id=doc_id)
-        return news.parse(options=self.create_parse_options(),
-                          frame_variant_collection=frame_variant_collection)
+        return TextParser.parse_news(news=news,
+                                     parse_options=self.create_parse_options())