+++ release_notes.md
@@ -1,6 +1,8 @@
 ## v0.20.5-rc
 
 Collection of parsed news become dispatched from text opinions collection.
+* **News parsing now is assumed to be performed using `TextParser.parse(news, options)` call. Related refactoring.**
+    * Stemmer application from `RuAttitiudes` parser has been removed.
 * Removed dependency from `RelatedParsedNewCollection` in TextOpinionCollection.
 * Labeling now separated from LinkedTextOpinion collection.
 * `ParsedText` class has been refactored, removed unused methods. Keep tokens has been discarded.